refactor: move convertProtoRecord from S2S module to shared proto.ts
Keeps the S2S transport's export surface private. The function now lives
alongside related helpers (decodeProtoReadBatch, fromProtoSequencedRecord)
in proto.ts, imported by both S2S and tests.
