+++ packages/streamstore/src/lib/stream/transport/proto.ts
@@ -2,7 +2,7 @@ import { S2Error } from "../../../error.js";
 import type * as API from "../../../generated/index.js";
 import * as Proto from "../../../generated/proto/s2.js";
 import type * as Types from "../../../types.js";
-import type { AppendRecord, ReadBatch } from "../types.js";
+import type { AppendRecord, ReadBatch, ReadRecord } from "../types.js";
 
 const textEncoder = new TextEncoder();
 
@@ -89,6 +89,63 @@ const fromProtoSequencedRecord = (
 		body: record.body,
 	};
 };
+
+/**
+ * Convert a raw protobuf SequencedRecord to the requested ReadRecord format.
+ * Used by the S2S transport for record conversion.
+ */
+export function convertProtoRecord<
+	Format extends "string" | "bytes" = "string",
+>(
+	record: {
+		seqNum?: bigint;
+		timestamp?: bigint;
+		headers?: Array<{ name?: Uint8Array; value?: Uint8Array }>;
+		body?: Uint8Array;
+	},
+	format: Format,
+	textDecoder: TextDecoder = new TextDecoder(),
+): ReadRecord<Format> {
+	if (format === "bytes") {
+		return {
+			seq_num: bigintToSafeNumber(
+				record.seqNum ?? 0n,
+				"SequencedRecord.seqNum",
+			),
+			timestamp: bigintToSafeNumber(
+				record.timestamp ?? 0n,
+				"SequencedRecord.timestamp",
+			),
+			headers: record.headers?.map(
+				(h) =>
+					[h.name ?? new Uint8Array(), h.value ?? new Uint8Array()] as [
+						Uint8Array,
+						Uint8Array,
+					],
+			),
+			body: record.body,
+		} as ReadRecord<Format>;
+	}
+	const headerEntries = record.headers?.map(
+		(h) =>
+			[
+				h.name ? textDecoder.decode(h.name) : "",
+				h.value ? textDecoder.decode(h.value) : "",
+			] as [string, string],
+	);
+	return {
+		seq_num: bigintToSafeNumber(
+			record.seqNum ?? 0n,
+			"SequencedRecord.seqNum",
+		),
+		timestamp: bigintToSafeNumber(
+			record.timestamp ?? 0n,
+			"SequencedRecord.timestamp",
+		),
+		headers: headerEntries,
+		body: record.body ? textDecoder.decode(record.body) : undefined,
+	} as ReadRecord<Format>;
+}
 
 export const buildProtoAppendInput = (
 	input: Types.AppendInput,
